pyproject.toml
[tool.poetry]
name = "flamingo"
version = "0.0.1"
description = "Flamingo"
authors = ["Joao Daher <joao@daher.dev>"]
packages = [
    { include = "flamingo" },
]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
sanic-rest = ">=1.7.0"
gcp-pilot = { version = "*", extras = ['datastore', 'build', 'storage', 'pubsub', 'dns']}
python-slugify = "*"
PyGithub = "*"
cryptography = "*"
pylint = { version = "*", optional = true}
coverage = { version = "*", optional = true}

[tool.poetry.extras]
dev-kit = ["pylint", "coverage"]

[tool.poetry.dev-dependencies]
pylint = "*"
coverage = "*"

[tool.black]
line-length = 120
target-version = ['py310']


[tool.pylint.messages_control]
disable = "C0330, C0326, C0114, C0115, C0116, W0212, W0511, W0613, W1203, W0221, R0901, R0903, R0913, R0201, R0801, R0401, R0902, R0914, E1101, R0912, C0204"

[tool.pylint.format]
max-line-length = "120"

[tool.pylint.basic]
good-names= "i, j, k, m, n, dt, db, id, pk, tz, qs, ex, _"

[tool.coverage.run]
branch = true
source = ["flamingo"]

[tool.pytest.ini_options]
addopts = "--cov"


[build-system]
requires = ["poetry"]
build-backend = "poetry.masonry.api"
